// src/types/TrackContainers.ts
export interface TrackContainer {
    devid: string;
    devtype: string;
    containerno: string;
    datatime: string;
    lng: string;
    lat: string;
    loctime: string;
    flag: string;
    batt: string;
    temp: string;
    staticT: string;
    description: string | null;
    type: string | null;
    is_empty: string;
    remark: string | null;
    alng: string;
    alat: string;
    inchina: string;
    address: string;
    key: string;
    coolerstate: string;
    reefer: {
        coolertype: string;
        coolertime: string;
        workmode: string;
        Boxnum: string;
        settemp: string;
        sstemp: string;
        rstemp: string;
        failcode: string;
        voltage: string;
        elecurrent: string;
        worktime: string;
        ambtemp: string;
        zftemp: string;
        colddoubtertmp: string;
        humidity: string;
        devid: string;
    };
    [key: string]: any;
};